requirements.txt
boto3==1.26.118
botocore==1.29.118
cachelib==0.9.0
click==8.1.3
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-bootstrap-templates==1.0.8
dash-core-components==2.0.0
dash-extensions==0.1.13
dash-html-components==2.0.0
dash-leaflet==0.1.23
dash-renderer==1.9.1
dash-table==5.0.0
EditorConfig==0.12.3
Flask==2.2.3
Flask-Caching==2.0.2
geobuf==1.1.1
importlib-metadata==6.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsbeautifier==1.14.7
MarkupSafe==2.1.2
more-itertools==9.1.0
numpy==1.24.2
packaging==23.1
pandas==2.0.0
plotly==5.14.1
protobuf==4.22.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
tenacity==8.2.2
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
requests
pyyaml